emailproj/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
django-rest-swagger==2.2.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
pip==9.0.3
pytz==2018.5
requests==2.19.1
setuptools==39.0.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
psycopg2==2.7.5
